refactor[collection]: group index related methods into 'IndexableTrait'
